refactor: move flags module to drive
