+++ drive/tests/query_tests.rs
@@ -3,8 +3,8 @@ use rand::seq::SliceRandom;
 use rand::{Rng, SeedableRng};
 use rs_drive::common;
 use rs_drive::common::{cbor_inner_bytes_value, setup_contract};
-use rs_drive::contract::flags::StorageFlags;
 use rs_drive::contract::{Contract, Document};
+use rs_drive::drive::flags::StorageFlags;
 use rs_drive::drive::object_size_info::DocumentAndContractInfo;
 use rs_drive::drive::object_size_info::DocumentInfo::DocumentAndSerialization;
 use rs_drive::drive::Drive;
